refactor(article-body): move any body related to new component
